requirements.txt
beautifulsoup4==4.12.2
feedparser==6.0.10
notion-client==2.0.0
python-dotenv==1.0.0
PyPDF2==3.0.1
requests==2.31.0 